+++ src/app/app-router.js
@@ -2,7 +2,8 @@ import Vue from 'vue';
 import VueRouter from 'vue-router';
 import { authRoutes } from './auth';
 import { courseRoutes } from './courses';
-import { AppProfile, AppWelcome ,AppNotFound } from './core/components';
+import { AppWelcome, AppNotFound } from './core/components';
+import { AppProfile } from './user/components';
 
 Vue.use(VueRouter);
 
@@ -10,7 +11,7 @@ const appRoutes = [
   {
     path: '/',
     name: 'welcome',
-    component: AppWelcome,
+    component: AppWelcome
   },
   {
     path: '/profile',
