+++ src/app/core/components/index.js
@@ -2,5 +2,4 @@ export { default as AppFooter } from './Footer.vue';
 export { default as AppNavbar } from './Navbar.vue';
 export { default as AppNotFound } from './NotFound.vue';
 export { default as AppWelcome } from './Welcome.vue';
-export { default as AppProfile } from './Profile.vue';
 export { default as AppSidenav } from './Sidenav.vue';
