+++ src/app/user/components/index.js
@@ -1,0 +1,2 @@
+export { default as AppProfile } from './Profile.vue';
+export { default as AppEditFrom } from './EditForm.vue';
